--- samples/IndexManagement/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "cosmosdb-index-management",
-  "version": "0.0.0",
-  "private": true,
-  "description": "A Sample to explain the many ways to set indexes on Azure Cosmos DB",
-  "scripts": {
-    "start": "node app.js"
-  },
-  "dependencies": {}
-}
