+++ samples/package.json
@@ -10,6 +10,7 @@
     "UserManagement": "npx ts-node ./UserManagement",
     "ServerSideScripts": "npx ts-node ./ServerSideScripts",
     "ItemManagement": "npx ts-node ./ItemManagement",
-    "DatabaseManagement": "npx ts-node ./DatabaseManagement"
+    "DatabaseManagement": "npx ts-node ./DatabaseManagement",
+    "IndexeManagement": "npx ts-node ./IndexeManagement"
   }
 }
